Scintilla: Fixed bug which caused unnecessary redraws.
I submitted a bug report to Neil and the mailing list.
Hopefully, the put it into the central code repo.

This one makes our spell checker much faster, since much less redraws are necessary.
Using the new performance timer I found out, the the actual spell checking
is neglectable. Scintillas redrawing uses the majority of time.

If Scintilla would have a single function to replace the indicators of a line,
that would greatly speed up the spell checking.

And to get a better user experience regarding spell checking:
We could refrain from checking the token under the cursor. But that is difficult.
We would need to post a message to ourselves to re-check that line later.
But currently we end up having words half-underlined, since they have been checked
before they were completed by the user.